chore: move middleware authentication into the middleware package
