+++ internal/middleware/authentication.go
@@ -1,4 +1,4 @@
-package handler
+package middleware
 
 import (
 	"crypto/rsa"
