+++ pkg/user/router.go
@@ -1,11 +1,11 @@
 package user
 
 import (
-	"github.com/dhis2-sre/im-manager/internal/handler"
+	"github.com/dhis2-sre/im-manager/internal/middleware"
 	"github.com/gin-gonic/gin"
 )
 
-func Routes(r *gin.Engine, authenticationMiddleware handler.AuthenticationMiddleware, handler Handler) {
+func Routes(r *gin.Engine, authenticationMiddleware middleware.AuthenticationMiddleware, handler Handler) {
 	r.POST("/users", handler.SignUp)
 	r.POST("/refresh", handler.RefreshToken)
